requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.1
attrs==25.3.0
beautifulsoup4==4.13.3
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorlog==6.9.0
dependency-groups==1.3.0
distlib==0.3.9
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.1
fuzzywuzzy==0.18.0
geographiclib==2.0
geopy==2.4.1
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
Levenshtein==0.27.1
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
npx==0.1.6
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
openai-whisper==20240930
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic_core==2.33.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytz==2025.2
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
sympy==1.13.1
tabula-py==2.10.0
threadpoolctl==3.6.0
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.3
